split server
